scripts/qflow/requirements.txt
rich>=10.0.0
pyyaml>=5.4.1
pathlib>=1.0.1
